requirements.txt
certifi==2019.11.28
chardet==3.0.4
docker==4.2.0
icalendar==4.0.4
idna==2.8
paho-mqtt==1.5.0
python-dateutil==2.8.1
python-dotenv==0.10.5
pytz==2019.3
requests==2.22.0
six==1.14.0
urllib3==1.25.7
websocket-client==0.57.0
